+++ std-bits/table/src/main/java/org/enso/table/data/table/problems/Problem.java
@@ -1,0 +1,8 @@
+package org.enso.table.data.table.problems;
+
+import java.util.List;
+
+public interface Problem {
+    String getMessage();
+}
+